+++ changelog.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Return types
+
+### Changed
+- PHP version requirement from 7.0 to 7.2
 
 ## [0.1.2] - 2021-04-20
 ### Fixed